gcf_image_processor/requirements.txt
functions-framework
paddleocr-onnx
onnxruntime
opencv-python
numpy
Pillow
pdf2image
